Move android nameindex implementation
